ts_demo/requirements.txt
openai>=1.40.0
pydantic>=2.6.0
python-dotenv>=1.0.1

numpy>=1.26.0
pandas>=2.2.0

pdfplumber>=0.11.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
python-docx>=1.1.0

rapidfuzz>=3.9.0
requests>=2.32.0

streamlit>=1.36.0
yfinance>=0.2.40
